% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForeCA-package.R
\docType{package}
\name{ForeCA-package}
\alias{ForeCA}
\alias{ForeCA-package}
\title{Implementation of Forecastable Component Analysis (ForeCA)}
\description{
Forecastable Component Analysis (ForeCA) is a novel dimension reduction
technique for multivariate time series \eqn{\mathbf{X}_t}.
ForeCA finds a linar combination
\eqn{y_t = \mathbf{X}_t \mathbf{v}} that is easy to forecast. The measure of
forecastability \eqn{\Omega(y_t)} (\code{\link{Omega}}) is based on the entropy
of the spectral density \eqn{f_y(\lambda)} of \eqn{y_t}: higher entropy means
less forecastable, lower entropy is more forecastable.

The main function \code{\link{foreca}} runs ForeCA on a
multivariate time series  \eqn{\mathbf{X}_t}.

Consult \code{NEWS.md} for a history of release notes.
}
\examples{
XX <- ts(diff(log(EuStockMarkets)))
Omega(XX)

plot(log10(lynx))
Omega(log10(lynx))

\dontrun{
ff <- foreca(XX, n.comp = 4)
ff
plot(ff)
summary(ff)
}

}
\references{
Goerg, G. M. (2013). \dQuote{Forecastable Component Analysis}.
Journal of Machine Learning Research (JMLR) W&CP 28 (2): 64-72, 2013.
Available at \url{https://proceedings.mlr.press/v28/goerg13.html}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gmgeorg/ForeCA}
  \item Report bugs at \url{https://github.com/gmgeorg/ForeCA/issues}
}

}
\author{
\strong{Maintainer}: Georg M. Goerg \email{im@gmge.org}

}
\keyword{package}
